Uint to uint8

Approved-by: zhangyaning <u2@cryptape.com>
Approved-by: Zhiwei <rink1969@cryptape.com>

+++ scripts/contracts/src/interfaces/sys_config.sol
@@ -51,6 +51,6 @@ interface ISysConfig {
     /// @notice The token information
     function getTokenInfo() external view returns (string, string, string);
 
-    function getEconomicalModel() external view returns (uint);
+    function getEconomicalModel() external view returns (uint8);
 
 }
